Further validate project title
Ensure there are no numbers in the title, the kebab-case route binding resolution does not properly format with numbers.

There should be no need to include any numbers in a project title. This is why it makes most sense to remove the ability to include them rather than modify the route binding logic.